### 2020-11-27 #2870 @whedon
[PRE REVIEW]: bcdata: An R package for searching & retrieving data from the B.C. Data Catalogue
**Submitting author:** @boshek (<a href="http://orcid.org/0000-0002-9270-7884">Samuel Albers</a>)
**Repository:** <a href="https://github.com/bcgov/bcdata" target ="_blank">https://github.com/bcgov/bcdata</a>
**Version:** 0.2.1
