### 2020-11-27 #2870 @arfon
@boshek - as [per our submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), could you please make sure your paper has a section headed 'Statement of need' describing the research need and purpose of this software for a non-specialist audience?
