### 2020-12-04 4f12b15
Only show progress bar when interactive
