+++ dev/pkgdown.yml
@@ -1,11 +1,11 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9000
-pkgdown_sha: e5847476d79f08b4cdca0ed5cc55bf55fc18a161
+pkgdown_sha: 6094ac3dd80a1f3f1d4e2f47d0ed4716d51362b0
 articles:
   bcdata: bcdata.html
   efficiently-query-spatial-data-in-the-bc-data-catalogue: efficiently-query-spatial-data-in-the-bc-data-catalogue.html
   explore-silviculture-data-using-bcdata: explore-silviculture-data-using-bcdata.html
-last_built: 2020-11-04T19:11Z
+last_built: 2020-12-04T23:33Z
 urls:
   reference: https://bcgov.github.io/bcdata/reference
   article: https://bcgov.github.io/bcdata/articles
