### 2020-12-13 #2870 @arfon
:wave: @boshek - could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?
