### 2020-12-20 #2870 @arfon
:wave: @MilesMcBain & @tweed1e - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission to be reviewed here is _bcdata: An R package for searching & retrieving data from the B.C. Data Catalogue_.

- The software: https://github.com/bcgov/bcdata
- The paper: [10.21105.joss.02870.pdf](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02870/joss.02870/10.21105.joss.02870.pdf)
