### 2020-12-24 #2927 @arfon
@tweed1e, @MilesMcBain - This is the review thread for the paper. All of our communications will happen here from now on.

Our bot Whedon will send an automated reminder in a couple of weeks to see how you're getting along.
