### 2020-12-29 #2927 @tweed1e
> I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

A potential COI is that I work for the Canadian federal government and this package was developed by the BC provincial government. But I've never worked with the authors of the pkg. @arfon can you waive that COI?
