### 2021-01-01 #2927 @arfon
> A potential COI is that I work for the Canadian federal government and this package was developed by the BC provincial government. But I've never worked with the authors of the pkg. @arfon can you waive that COI?

Yes, this is fine. Thanks for disclosing this - please proceed.
