### 2021-01-04 #2927 @arfon
> Sorry @arfon I can't check boxes in the lists, probably because the collaborator invite expired?

Yes, I think so. Please try the invite link again as I've now re-invited you.
