### 2021-01-10 #2927 @MilesMcBain
Hi @whedon and @afron Things are progressing smoothly so far. I'm part-way through verifying the functionality. I don't anticipate this software generating major issues or requiring extra time based on what I have seen so far. :+1:
