### 2021-01-10 bcgov/bcdata#239 @MilesMcBain
JOSS Review: Usability comments
Hi,

Congratulations on putting together a very polished and usable interface to your province's open data!

I have a some usability feedback arising from my in-progress JOSS review, and there will probably be more, so let's keep this thread open.

Unless otherwise stated I don't believe any of these issues are blockers for JOSS acceptance.

UF1: Is it possible to either display the file size of resources associated with the record, the proportion of file size downloaded with the progress bar, or prompt for confirmation when a download is about to be initiated above some threshold?

For example:

`bcdc_get_data("e2dadc60-292f-4d98-b42b-56ca9e4fe694")` lets me know that "This record requires pagination to complete the request." and displays a progress bar while download completes. But it's hard to know if I have made a reasonable request or if it will return in a reasonable amount of time since I do not know how large the data are, or what portion of the file size the progress represents.

UF2: It would be convenient to be able to fetch the accompanying metadata for some data with a `bcdc_` function.

Consider the Fire location data: https://catalogue.data.gov.bc.ca/dataset/fire-incident-locations-historical, it has a helpful data description table under "Object Description". An example of how this could be useful: facilitate comparisons between the column data types chosen by R's parsers vs the native data types.

Edit: linking back to main review thread: https://github.com/openjournals/joss-reviews/issues/2927
