### 2021-01-12 bcgov/bcdata#240 @ateucher
Provide information on filesize/expected download time
From @MilesMcbain JOSS review feedback: #239

UF1: Is it possible to either display the file size of resources associated with the record, the proportion of file size downloaded with the progress bar, or prompt for confirmation when a download is about to be initiated above some threshold?

For example:

bcdc_get_data("e2dadc60-292f-4d98-b42b-56ca9e4fe694") lets me know that "This record requires pagination to complete the request." and displays a progress bar while download completes. But it's hard to know if I have made a reasonable request or if it will return in a reasonable amount of time since I do not know how large the data are, or what portion of the file size the progress represents.
