### 2021-01-12 bcgov/bcdata#239 @ateucher
Hi @MilesMcBain - great to have you reviewing the paper, and thanks so much for making the time and effort to post comments here as well.

For your first point regarding file sizes/expected download time, that is a really good observation and adding this kind of information would help a lot. I've opened a separate issue for us to track that (#240).

UF2:  We do have `bcdc_get_record()` and `bcdc_describe_feature()`:

``` r
library(bcdata)
#>
#> Attaching package: 'bcdata'
#> The following object is masked from 'package:stats':
#>
#>     filter

bcdc_get_record("fire-incident-locations-historical")
#> Authorizing with your stored API key
#> Warning: It is advised to use the permanent id ('e2dadc60-292f-4d98-b42b-56ca9e4fe694') rather than the name of the record ('fire-incident-locations-historical') to guard against future name changes.
#> B.C. Data Catalogue Record: Fire Incident Locations - Historical
#> Name: fire-incident-locations-historical (ID:
#>  e2dadc60-292f-4d98-b42b-56ca9e4fe694)
#> Permalink:
#>  https://catalogue.data.gov.bc.ca/dataset/e2dadc60-292f-4d98-b42b-56ca9e4fe694
#> Sector: Natural Resources
#> Licence: Open Government Licence - British Columbia
#> Type: Geographic
#> Last Updated: 2021-01-01
#> Description: Wildfire historic incident point locations for all fire
#>  seasons before the current season.  Supplied through various sources.
#>  Not to be used for legal purposes.  This data includes all incidents
#>  tracked by BC Wildfire Service, ie. actual fires, suspected fires,
#>  nusiance fires, smoke chases, etc.  On April 1 of each year this layer
#>  is updated with the previous fire season's data
#> Resources: (3)
#> # A tibble: 3 x 8
#>   name     url        id      format ext   package_id  location bcdata_available
#>   <chr>    <chr>      <chr>   <chr>  <chr> <chr>       <chr>    <lgl>
#> 1 WMS get… https://o… eb2fff… wms    ""    e2dadc60-2… bcgeogr… TRUE
#> 2 KML Net… http://op… 012b8c… kml    "kml" e2dadc60-2… bcgeogr… FALSE
#> 3 BC Geog… https://c… 6db589… other  ""    e2dadc60-2… bcgeogr… FALSE
#> You can access the 'Resources' data frame using bcdc_tidy_resources()

bcdc_describe_feature("fire-incident-locations-historical")
#> Authorizing with your stored API key
#> # A tibble: 18 x 4
#>    col_name               sticky remote_col_type       local_col_type
#>    <chr>                  <lgl>  <chr>                 <chr>
#>  1 id                     FALSE  xsd:string            character
#>  2 FIRE_NUMBER            FALSE  xsd:string            character
#>  3 FIRE_YEAR              FALSE  xsd:decimal           numeric
#>  4 IGNITION_DATE          TRUE   xsd:date              date
#>  5 FIRE_CAUSE             TRUE   xsd:string            character
#>  6 FIRE_LABEL             TRUE   xsd:string            character
#>  7 FIRE_CENTRE            FALSE  xsd:decimal           numeric
#>  8 ZONE                   FALSE  xsd:decimal           numeric
#>  9 FIRE_ID                FALSE  xsd:decimal           numeric
#> 10 FIRE_TYPE              TRUE   xsd:string            character
#> 11 GEOGRAPHIC_DESCRIPTION TRUE   xsd:string            character
#> 12 LATITUDE               TRUE   xsd:decimal           numeric
#> 13 LONGITUDE              TRUE   xsd:decimal           numeric
#> 14 CURRENT_SIZE           TRUE   xsd:decimal           numeric
#> 15 FEATURE_CODE           TRUE   xsd:string            character
#> 16 SHAPE                  TRUE   gml:PointPropertyType sfc geometry
#> 17 OBJECTID               FALSE  xsd:decimal           numeric
#> 18 SE_ANNO_CAD_DATA       TRUE   xsd:hexBinary         numeric
```

However, I see now that doesn't contain as much information as the "Object Description". That information is available via the catalogue API, so shouldn't be difficult to add and would be helpful, as you say.
