### 2021-01-12 bcgov/bcdata#241 @ateucher
Add "Object Description" from catalogue API to bcdc_get_record
From @MilesMcbain JOSS review feedback: #239

UF2: It would be convenient to be able to fetch the accompanying metadata for some data with a bcdc_ function.

Consider the Fire location data: https://catalogue.data.gov.bc.ca/dataset/fire-incident-locations-historical, it has a helpful data description table under "Object Description". An example of how this could be useful: facilitate comparisons between the column data types chosen by R's parsers vs the native data types.
