### 2021-01-12 bcgov/bcdata#241 @ateucher
I think it would be best to add it to `bcdc_get_record` vs `bcdc_describe_feature` to avoid mixing information from the two APIs (ckan/WFS)
