### 2021-01-12 bcgov/bcdata#239 @MilesMcBain
UF3: Should you state explicitly in the documentation that the format of a resource needs to be `wms` to use `bcdc_query_geodata`?

The help for `bcdc_query_geodata` says:

> Queries features from the B.C. Web Service. The data must be available as a Web Service. ...

And the closest statement I have found is in the `Querying Spatial Data with bcdata` vignette:

>  This data is the boundary of each school district. The key information in this metadata is that the layer is a `**WFS request (Spatial Data)**`---which means it is available through a [Web Service](https://www2.gov.bc.ca/gov/content?id=95D78D544B244F34B89223EF069DF74E). From this we know we can make use of `bcdc_query_geodata`.

I am 95% sure this corresponds to the `wms` format based on the content of the `Web Service` link. If true, this could be made more explicit by changing that para to something like:

> This data is the boundary of each school district. The key information in this metadata is that the available resource has `format` `**wms**`---which means it is available through a [Web Maps Service](https://www2.gov.bc.ca/gov/content?id=95D78D544B244F34B89223EF069DF74E). From this we know we can make use of `bcdc_query_geodata`.
