### 2021-01-12 bcgov/bcdata#239 @ateucher
Thanks for this as well - we've struggled to keep those terms consistent across the documentation (and what level of abstraction to expose the user to - i.e., do/should they care what the difference is between WFS and WMS?)
