### 2021-01-16 bcgov/bcdata#239 @MilesMcBain
UF4: Would it be worth implementing `distinct.bcdc_promise`?

This could resolve the issue mentioned in the spatial query vignette:

>For example, there is currently no way to ask the catalogue for all possible unique values of `SCHOOL_DISTRICT_NAME`. Is that case the data will need to be brought into R and unique values will need to be determined there.

The user could do:

```
bcdc_query_geodata("78ec5279-4534-49a1-97e8-9d315936f08b") %>%
  distinct(SCHOOL_DISTRICT_NAME) %>%
  collect()
```

The current option involving pulling the entire `SCHOOL_DISTRICT_NAME` down to the local session takes a while on my connection, so I think this could be a useful one for tuning queries.
