### 2021-01-17 bcgov/bcdata#242 @MilesMcBain
JOSS Review: Software Citations
I was just looking at the list of citations in the paper and I think there's an opportunity to cite some more software.

I do think `sf` should be cited when you first refer to an 'sf object' being returned - you've already hyperlinked the `sf` repository.

The others that I am less firm on but are worth considering are `readr` and `readxl`. They're doing some heavy lifting, and they form part of your API since you expose parts of their functionality to your users via the `...` of `bcdc_get_data`.

Edit: linking this to main review thread: https://github.com/openjournals/joss-reviews/issues/2927
