### 2021-01-18 bcgov/bcdata#239 @ateucher
Thanks again @MilesMcBain - it is indeed unfortunate that there is no `DISTINCT` or analogue in the WFS... not sure there's a good way around that.

Those are excellent suggestions around the conversion to bounding box - I again opened a separate issue where we'll tackle that.
