### 2021-01-19 bcgov/bcdata#244 @MilesMcBain
[JOSS Review]: Statement of Need
I haven't yet read anything in the documentation or JOSS paper draft, that I would consider a statement of need. This is a blocker to acceptance since it is specifically called out in the review checklist.

This is what I read in the paper:

> Statement of need
  bcdata is an R package that enables programmatic access, using the R programming language,
to data sets housed in the British Columbia (B.C.) Data Catalogue. The package connects
the diverse array of mapping, modelling and data processing capabilities of the R ecosystem
to the hundreds of open-licensed data sets publicly available in the B.C. Data Catalogue.

It's more of a capability statement. My understanding of this statement is that it is supposed to articulate why this software needed to be developed, and who for.

I think it should be pretty easy to do this. Some angles that come to mind regarding need:

* Reproducibility. Scripts can easilly refer to bcdata sources by static GUID. More likely to work in future. Data lineage can be made completely clear.
* Convenience. Esp for working with spatial data. Helps avoid large downloads. Automatically parses data into readily usable form with correct spatial metadata. Helps non-GIS-experts achieve spatial analysis tasks.

You'll probably have others.

Audience is a bit easier: people analysing public data with R in BC. They can be beginner to intermediate useRs since `bcdata` is flattening the learning curve a bit.

Link to JOSS review thread:  https://github.com/openjournals/joss-reviews/issues/2927
