### 2021-01-20 bcgov/bcdata#245 @MilesMcBain
[JOSS Review]: State of the field
A review checklist item for the paper is that it discusses the 'state of the field'. My understanding is that is where you draw comparisons between your software and other software people might use to achieve similar outcomes.

In your case you might mention:
  * [ckanr](https://github.com/ropensci/ckanr) for talking to CKAN file stores
  * [ows4R](https://github.com/eblondel/ows4R) or [rwfs](https://github.com/rOpenGov/rwfs) for querying geospatial data via WFS

As an example you could talk about how you unify these concerns in one tool, abstract away low level protocol details, and utilise a functional programming model that is compatible with `dplyr` and the wider `tidyverse` suite, rather than the OOP style of these packages.

Another thing you might consider is comparisons to R packages that are designed to connect to other open data portals. For example:
  * [opendatatoronto](https://github.com/sharlagelfand/opendatatoronto/blob/main/R/packages.R) is another package that wraps a CKAN file store.
  * [BARIS](https://github.com/feddelegrand7/BARIS) seems similar in its workflow - though not for spatial data
  * More to be found in [this twitter thread](https://twitter.com/MilesMcBain/status/1351483619596877824)? and at https://github.com/ropensci/opendata

I am not suggesting an exhaustive review. More like here are a few examples that show there is precedent for this work, they have these features the same as ours (search, fetch etc) but not the `dbplyr` integration for large spatial data, which definitely seems to be novel and useful.

edit: link main review thread: https://github.com/openjournals/joss-reviews/issues/2927
