### 2021-01-20 bcgov/bcdata#246 @MilesMcBain
[JOSS Review]: Community Guidelines - Seeking Support
An element of the Community Guidelines documentation checklist item is that you should provide clear guidelines for third parties who want to seek support.

In your README I see:

> Getting Help or Reporting an Issue
  To report bugs/issues/feature requests, please file an issue.

Which is not quite clear direction as to what someone should do if they're in the "Getting Help" category.

edit: link review thread: https://github.com/openjournals/joss-reviews/issues/2927
