### 2021-01-20 #2927 @MilesMcBain
I've finished my first pass review.

This is a really great package, potentially best in class.

The `dbplyr` backend for the `WFS` API is a genuinely novel development for these types of packages, and I hope it is copied by other similar packages.

I also appreciate from an end user's perspective the attention to detail in the print methods, messages, errors, and warnings. It's quite chatty in a reassuring way, and I think it lowers the level of R experience required for good results significantly.

The documentation is high quality, especially the three vignettes, which do a good job of unveiling features without overwhelming.

There are a few issues to resolve but all fairly minor. Congratulations to the authors!

Current blockers to acceptance are:
* https://github.com/bcgov/bcdata/issues/244
* https://github.com/bcgov/bcdata/issues/246
* https://github.com/bcgov/bcdata/issues/245
* https://github.com/bcgov/bcdata/issues/242

And I would strongly recommend taking some action to rectify usability feedback item 5 (UF5):
* https://github.com/bcgov/bcdata/issues/243
