### 2021-01-20 bcgov/bcdata#246 @boshek
This is a good point. For a project like this, a stack overflow tag seems like overfill. Do we think referring the user to vignettes and pkgdown site first for help, _then_ telling them to open an issue for help would be sufficient?
