### 2021-01-21 bcgov/bcdata#246 @MilesMcBain
A thing I've seen done is using the GitHub issue templates to create checklist of things you expect the user to have done before seeking support. You can have separate templates for bugs vs feature requests vs support if you want.

Here's the template from https://github.com/rstudio/rmarkdown/issues/new

```

---

By filing an issue to this repo, I promise that

- [ ] I have fully read the issue guide at https://yihui.org/issue/.
- [ ] I have provided the necessary information about my issue.
    - If I'm asking a question, I have already asked it on Stack Overflow or RStudio Community, waited for at least 24 hours, and included a link to my question there.
    - If I'm filing a bug report, I have included a minimal, self-contained, and reproducible example, and have also included `xfun::session_info('rmarkdown')`. I have upgraded all my packages to their latest versions (e.g., R, RStudio, and R packages), and also tried the development version: `remotes::install_github('rstudio/rmarkdown')`.
    - If I have posted the same issue elsewhere, I have also mentioned it in this issue.
- [ ] I have learned the Github Markdown syntax, and formatted my issue correctly.

I understand that my issue may be closed if I don't fulfill my promises.
```

You could also use this to set expectations as to the level of support provided. E.g. you could say support is best effort, but responses to support requests may take some time etc. quicker responses may be had by asking on Stack Overflow, or the #rstats hashtag on Twitter, or any other relevant communities.
