### 2021-01-21 bcgov/bcdata#246 @stephhazlitt
Oh I do like the checklist ✅ proposal. Tx @MilesMcBain.
