### 2021-01-21 48c66f1
add groups function. Closes #234
