+++ man/bcdc_group.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/bcdc_search.R
+\name{bcdc_group}
+\alias{bcdc_group}
+\title{Retrieve all metadata of all data within a group}
+\usage{
+bcdc_group(group)
+}
+\arguments{
+\item{group}{Name of the group}
+}
+\description{
+Access a tibble of record id's by providing the group name. Groups can be view
+here: https://catalogue.data.gov.bc.ca/group
+}
+\examples{
+\donttest{
+bcdc_group('environmental-reporting-bc')
+}
+}
