+++ NAMESPACE
@@ -20,6 +20,7 @@ S3method(collect,bcdc_promise)
 S3method(filter,bcdc_promise)
 S3method(head,bcdc_promise)
 S3method(mutate,bcdc_promise)
+S3method(print,bcdc_group)
 S3method(print,bcdc_promise)
 S3method(print,bcdc_query)
 S3method(print,bcdc_record)
@@ -48,6 +49,7 @@ export(bcdc_browse)
 export(bcdc_describe_feature)
 export(bcdc_get_data)
 export(bcdc_get_record)
+export(bcdc_group)
 export(bcdc_list)
 export(bcdc_options)
 export(bcdc_preview)
