### 2021-01-25 5392835
add 'groups' to facet search
