### 2021-01-25 e351ff6
add bcdc_list_groups
