+++ NAMESPACE
@@ -51,6 +51,7 @@ export(bcdc_get_data)
 export(bcdc_get_record)
 export(bcdc_group)
 export(bcdc_list)
+export(bcdc_list_groups)
 export(bcdc_options)
 export(bcdc_preview)
 export(bcdc_query_geodata)
