+++ man/bcdc_list_group_records.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/bcdc_search.R
-\name{bcdc_group}
-\alias{bcdc_group}
+\name{bcdc_list_group_records}
+\alias{bcdc_list_group_records}
 \title{Retrieve all metadata of all data within a group}
 \usage{
-bcdc_group(group)
+bcdc_list_group_records(group)
 }
 \arguments{
 \item{group}{Name of the group}
@@ -16,6 +16,6 @@ here: https://catalogue.data.gov.bc.ca/group or accessed directly from R using
 }
 \examples{
 \donttest{
-bcdc_group('environmental-reporting-bc')
+bcdc_list_group_records('environmental-reporting-bc')
 }
 }
