+++ NAMESPACE
@@ -49,8 +49,8 @@ export(bcdc_browse)
 export(bcdc_describe_feature)
 export(bcdc_get_data)
 export(bcdc_get_record)
-export(bcdc_group)
 export(bcdc_list)
+export(bcdc_list_group_records)
 export(bcdc_list_groups)
 export(bcdc_options)
 export(bcdc_preview)
