### 2021-01-26 857a48c
add new functions
