+++ _pkgdown.yml
@@ -15,6 +15,8 @@ reference:
       - '`bcdc_list`'
       - '`bcdc_get_record`'
       - '`bcdc_tidy_resources`'
+      - '`bcdc_list_groups`'
+      - '`bcdc_list_group_records`'
   - title: "Direct downloads of data sets"
     desc: >
       Directly download (or preview) data sets (resources) from the
