### 2021-02-01 bcgov/bcdata#248 @stephhazlitt
This is a great suggestion, thanks @tweed1e.
