### 2021-02-02 5f10a05
document groups fns in one place
