+++ man/bcdc_list_groups.Rd
@@ -1,17 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/bcdc_search.R
-\name{bcdc_list_groups}
-\alias{bcdc_list_groups}
-\title{List of all groups within the B.C. Data Catalogue}
-\usage{
-bcdc_list_groups()
-}
-\description{
-Returns a tibble of all current groups within the B.C. Data Catalogue.
-}
-\examples{
-\donttest{
-bcdc_list_groups()
-}
-
-}
