+++ man/bcdc_list_group_records.Rd
@@ -1,19 +1,26 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/bcdc_search.R
-\name{bcdc_list_group_records}
+\name{bcdc_list_groups}
+\alias{bcdc_list_groups}
 \alias{bcdc_list_group_records}
-\title{Retrieve all metadata of all data within a group}
+\title{Retrieve group information for B.C. Data Catalogue}
 \usage{
+bcdc_list_groups()
+
 bcdc_list_group_records(group)
 }
 \arguments{
 \item{group}{Name of the group}
 }
 \description{
-Returns a tibble of records by providing the group name. Groups can be view
-here: https://catalogue.data.gov.bc.ca/group or accessed directly from R using
-\href{`bcdc_list_groups`}{bcdc_list_groups}
+Returns a tibble of groups or records. Groups can be viewed here:
+https://catalogue.data.gov.bc.ca/group or accessed directly from R using \code{bcdc_list_groups}
 }
+\section{Functions}{
+\itemize{
+\item \code{bcdc_list_groups}: 
+}}
+
 \examples{
 \donttest{
 bcdc_list_group_records('environmental-reporting-bc')
