+++ dev/pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   bcdata: bcdata.html
   efficiently-query-spatial-data-in-the-bc-data-catalogue: efficiently-query-spatial-data-in-the-bc-data-catalogue.html
   explore-silviculture-data-using-bcdata: explore-silviculture-data-using-bcdata.html
-last_built: 2021-01-26T19:07Z
+last_built: 2021-02-02T22:39Z
 urls:
   reference: https://bcgov.github.io/bcdata/reference
   article: https://bcgov.github.io/bcdata/articles
