### 2021-03-01 ff11607
validate R6 Class fields have the same name
