### 2021-03-02 2830fb0
Add option for single page download size limit. Related to #250
