### 2021-03-02 cdb017f
bump min crul version to 1.1 due to Paginator API changes
