+++ tests/testthat/test-options.R
@@ -26,3 +26,13 @@ test_that("bcdata.chunk_limit",{
   options(bcdata.chunk_limit = original_value$bcdata.chunk_limit)
 
 })
+
+test_that("bcdata.single_download_limit", {
+  withr::local_options(list(bcdata.single_download_limit = 1))
+  expect_message(
+    bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8', resource =
+                    '4d0377d9-e8a1-429b-824f-0ce8f363512c'),
+    "pagination"
+  )
+
+})
