### 2021-03-02 346b5e5
Use withr for local options
