+++ .github/workflows/cmd-check.yaml
@@ -16,12 +16,13 @@ jobs:
       fail-fast: false
       matrix:
         config:
-        - { os: windows-latest, r: 'release', args: "'--no-manual'"}
-        - { os: macOS-latest,   r: 'release', args: "'--as-cran'"}
-        - { os: ubuntu-latest,   r: '3.5',     args: "'--no-manual'"}
-        - { os: ubuntu-latest,   r: 'oldrel',  args: "'--no-manual'"}
-        - { os: ubuntu-latest,   r: 'release', args: "c('--no-manual', '--as-cran')"}
-        - { os: ubuntu-latest,   r: 'devel',   args: "c('--no-manual', '--as-cran')"}
+        - { os: windows-latest,  r: 'release', args: "'--no-manual'"}
+        - { os: macOS-latest,    r: 'release', args: "'--no-manual'"}
+        - { os: macOS-latest,    r: 'oldrel',  args: "''"}
+        - { os: ubuntu-18.04,   r: '3.5',     rspm: "https://packagemanager.rstudio.com/cran/__linux__/bionic/latest", args: "'--no-manual'"}
+        - { os: ubuntu-18.04,   r: 'oldrel',  rspm: "https://packagemanager.rstudio.com/cran/__linux__/bionic/latest", args: "'--no-manual'"}
+        - { os: ubuntu-18.04,   r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/bionic/latest", args: "''"}
+        - { os: ubuntu-18.04,   r: 'devel',   rspm: "https://packagemanager.rstudio.com/cran/__linux__/bionic/latest", args: "'--no-manual'"}
 
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
@@ -71,7 +72,7 @@ jobs:
           sudo -s eval "$sysreqs"
 
           # install spatial dependencies
-          sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable
+          # sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable
           sudo apt update
           sudo apt install \
             libudunits2-dev \
