+++ .github/workflows/cmd-check.yaml
@@ -16,9 +16,9 @@ jobs:
       fail-fast: false
       matrix:
         config:
-        - { os: windows-latest,  r: 'release', args: "'--no-manual'"}
-        - { os: macOS-latest,    r: 'release', args: "'--no-manual'"}
-        - { os: macOS-latest,    r: 'oldrel',  args: "''"}
+        - { os: windows-latest, r: 'release', args: "'--no-manual'"}
+        - { os: macOS-latest,   r: 'release', args: "'--no-manual'"}
+        - { os: macOS-latest,   r: 'oldrel',  args: "''"}
         - { os: ubuntu-18.04,   r: '3.5',     rspm: "https://packagemanager.rstudio.com/cran/__linux__/bionic/latest", args: "'--no-manual'"}
         - { os: ubuntu-18.04,   r: 'oldrel',  rspm: "https://packagemanager.rstudio.com/cran/__linux__/bionic/latest", args: "'--no-manual'"}
         - { os: ubuntu-18.04,   r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/bionic/latest", args: "''"}
@@ -26,9 +26,10 @@ jobs:
 
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
+      RSPM: ${{ matrix.config.rspm }}
       CRAN: ${{ matrix.config.cran }}
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      cache-version: v5
+      cache-version: v6
 
     steps:
       - uses: actions/checkout@v2
