### 2021-03-11 961fb3e
update comments and NEWS
