+++ tests/testthat/test-get-data.R
@@ -165,8 +165,7 @@ test_that("bcdc_get_data handles sheet name specification", {
   skip_on_cran()
   expect_message(bcdc_get_data('8620ce82-4943-43c4-9932-40730a0255d6'), 'This .xlsx resource contains the following sheets:')
   expect_error(bcdc_get_data('8620ce82-4943-43c4-9932-40730a0255d6', sheet = "foo"), "Error: Sheet 'foo' not found")
-  out <- capture.output(bcdc_get_data('8620ce82-4943-43c4-9932-40730a0255d6', sheet = "Single Detached"), type = 'message')
+  out <- capture.output(bcdc_get_data('8620ce82-4943-43c4-9932-40730a0255d6', sheet = "Notes"), type = 'message')
   expect_false(any(grepl('This .xlsx resource contains the following sheets:', out)))
-
 })
 
