### 2021-03-11 621ee32
stop building manual on mac
