+++ .github/workflows/cmd-check.yaml
@@ -18,7 +18,7 @@ jobs:
         config:
         - { os: windows-latest, r: 'release', args: "'--no-manual'"}
         - { os: macOS-latest,   r: 'release', args: "'--no-manual'"}
-        - { os: macOS-latest,   r: 'oldrel',  args: "''"}
+        - { os: macOS-latest,   r: 'oldrel',  args: "'--no-manual'"}
         - { os: ubuntu-18.04,   r: '3.5',     rspm: "https://packagemanager.rstudio.com/cran/__linux__/bionic/latest", args: "'--no-manual'"}
         - { os: ubuntu-18.04,   r: 'oldrel',  rspm: "https://packagemanager.rstudio.com/cran/__linux__/bionic/latest", args: "'--no-manual'"}
         - { os: ubuntu-18.04,   r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/bionic/latest", args: "''"}
