### 2021-03-16 bcgov/bcdata#241 @acerickson
just here to endorse the addition of this feature - being able to get or join the data dictionary often tabulated under "Object Description".   The bcdc_describe_feature is great, but tells me nothing about the data represented by the cryptic variable names.
thanks! this is a great package!

![image](https://user-images.githubusercontent.com/43129048/111376048-3937bf00-865c-11eb-8309-3212368668a8.png)
