### 2021-03-17 a46bf78
skip an options test on CRAN
