### 2021-03-17 bcgov/bcdata#241 @acerickson
ah-ha, it's under 'details'!!
I was pulling the metadata using bcdc_get_record and looked under some of the collapsed lists but obviously missed that one. I figured that info was likely there somewhere! Maybe what all is needed is to add this step to the 'getting started' vignette to get data dictionary.
thanks @boshek Sam
