### 2021-03-17 05cfef1
smaller test; don't test specific values in wkt
