### 2021-03-17 bcgov/bcdata#241 @boshek
I think that a function could still be warranted here. I wonder if `bcdc_describe_record` might be a good function name. I am seeing that most records actually don't have object descriptions and that `$details` slot is empty. But even that is useful information.
