### 2021-03-25 d23247a
add fn to get download limit direct from wfs
