+++ R/utils.R
@@ -210,10 +210,10 @@ gml_types <- function(x) {
 
 get_record_warn_once <- function(...) {
   silence <- isTRUE(getOption("silence_named_get_record_warning"))
-  warned <- bcdata_env$named_get_record_warned
+  warned <- ._bcdataenv_$named_get_record_warned
   if (!silence && !warned) {
     warning(..., call. = FALSE)
-    assign("named_get_record_warned", TRUE, envir = bcdata_env)
+    assign("named_get_record_warned", TRUE, envir = ._bcdataenv_)
   }
 }
 
