### 2021-03-25 919333c
test if download limit option can be changed
