### 2021-03-26 e5502b8
Update tests/testthat/test-options.R
Co-authored-by: Andy Teucher <andy.teucher@gov.bc.ca>
