+++ tests/testthat/test-options.R
@@ -35,9 +35,10 @@ test_that("bcdata.single_download_limit", {
 })
 
 test_that("bcdata.single_download_limit can be changes",{
-  lt <- withr::local_options(list(bcdata.single_download_limit = 13),
-                       getOption("bcdata.single_download_limit"))
-  expect_equal(lt$bcdata.single_download_limit, 13)
+  withr::local_options(list(bcdata.single_download_limit = 13))
+                       
+  expect_equal(getOption("bcdata.single_download_limit"), 13)
+  expect_equal(bcdc_single_download_limit(), 13)
 })
 
 test_that("bcdc_single_download_limit returns a number",{
