### 2021-03-26 d856292
fix so that pkg load doesn't depend on internet access :scream:
