### 2021-03-26 fe1e472
assign option defaults to pkg env
