+++ tests/testthat/test-options.R
@@ -36,9 +36,7 @@ test_that("bcdata.single_download_limit", {
 
 test_that("bcdata.single_download_limit can be changes",{
   withr::local_options(list(bcdata.single_download_limit = 13))
-                       
   expect_equal(getOption("bcdata.single_download_limit"), 13)
-  expect_equal(bcdc_single_download_limit(), 13)
 })
 
 test_that("bcdc_single_download_limit returns a number",{
