### 2021-03-26 bcgov/bcdata#241 @ateucher
After discussion with @boshek  and @stephhazlitt, we decided to add the necessary fields from the `$details` slot to `bcdc_describe_feature`. Usability trumps under the hood design in this case.
