### 2021-03-26 bcgov/bcdata#245 @boshek
Probably worth mentioning this package: https://github.com/eblondel/ows4R
