### 2021-03-26 0da897c
Extract download limits directly from WFS (#256)
* add fn to get download limit direct from wfs

* set option on package load

* provide chunk limit directly

* remove hard coded dl limit

* rename pkg env

* fix typo

* 2nd typo

* fix typo #3

* test if download limit option can be changed

* Update tests/testthat/test-options.R

Co-authored-by: Andy Teucher <andy.teucher@gov.bc.ca>

* fix so that pkg load doesn't depend on internet access :scream:

* assign option defaults to pkg env

* collect default from pkg env

* clean and NEWS item

Co-authored-by: Andy Teucher <andy.teucher@gov.bc.ca
[…]
