+++ R/zzz.R
@@ -10,10 +10,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
 
-bcdata_env <- new.env(parent = emptyenv())
+._bcdataenv_ <- new.env(parent = emptyenv())
 
 .onLoad <- function(...) {
- assign("named_get_record_warned", FALSE, envir = bcdata_env) # nocov
+  ._bcdataenv_$named_get_record_warned <- FALSE # nocov
+  ._bcdataenv_$bcdata_dl_limit <- bcdc_single_download_limit() # nocov
+
 }
 
 # Define bcdc_sf as a subclass of sf so that it works
