+++ NEWS.md
@@ -1,3 +1,7 @@
+# bcdata 0.2.2.9000
+### IMPROVEMENTS
+- Setting the `bcdata.single_download_limit` limit dynamically from the getCapabilities endpoint. 
+
 # bcdata 0.2.2
 ### IMPROVEMENTS
 * Added `bcdc_list_groups` and `bcdc_list_group_records` to provide the ability to query on the group endpoint of the catalogue API. #234
