### 2021-03-31 b5e3947
add bcdc_check_size fn
