+++ NAMESPACE
@@ -58,6 +58,7 @@ export(bcdc_query_geodata)
 export(bcdc_read_functions)
 export(bcdc_search)
 export(bcdc_search_facets)
+export(bcdc_size_check)
 export(bcdc_tidy_resources)
 export(collect)
 export(filter)
