### 2021-03-31 d6fbd3f
add convenience cli fns
