+++ vignettes/efficiently-query-spatial-data-in-the-bc-data-catalogue.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Querying Spatial Data with bcdata"
-date: "2020-06-19"
+date: "2021-03-31"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Querying Spatial Data with bcdata}
@@ -47,29 +47,26 @@ Our first step is to extract the [school district polygons](https://catalog.data
 
 ```r
 bcdc_get_record("78ec5279-4534-49a1-97e8-9d315936f08b")
-#> B.C. Data Catalogue Record: School Districts of BC 
-#> 
-#> Name: school-districts-of-bc (ID: 78ec5279-4534-49a1-97e8-9d315936f08b )
-#> Permalink: https://catalogue.data.gov.bc.ca/dataset/78ec5279-4534-49a1-97e8-9d315936f08b
+#> B.C. Data Catalogue Record: School Districts of BC
+#> Name: school-districts-of-bc (ID: 78ec5279-4534-49a1-97e8-9d315936f08b)
+#> Permalink:
+#>  https://catalogue.data.gov.bc.ca/dataset/78ec5279-4534-49a1-97e8-9d315936f08b
 #> Sector: Education
 #> Licence: Open Government Licence - British Columbia
 #> Type: Geographic
-#> Last Updated: 2020-03-09 
-#> 
-#> Description: The School Districts dataset contains the spatial representation (polygon) of the
-#> current extent of the administrative areas defined under section 176(1) of the
-#> School Act for the purposes of preservation and promotion of the fundamental
-#> principle of local autonomy and control of public education at the public and
-#> governmental levels through locally elected school boards. 
-#> 
+#> Last Updated: 2020-11-03
+#> Description: The School Districts dataset contains the spatial representation
+#>  (polygon) of the current extent of the administrative areas defined under section
+#>  176(1) of the School Act for the purposes of preservation and promotion of the
+#>  fundamental principle of local autonomy and control of public education at the public
+#>  and governmental levels through locally elected school boards.
 #> Resources: (3)
 #> # A tibble: 3 x 8
-#>   name     url        id      format ext   package_id  location bcdata_available
-#>   <chr>    <chr>      <chr>   <chr>  <chr> <chr>       <chr>    <lgl>           
-#> 1 BC Geog… https://c… 529507… other  ""    78ec5279-4… bcgeogr… FALSE           
-#> 2 WMS get… https://o… 57d884… wms    ""    78ec5279-4… bcgeogr… TRUE            
-#> 3 KML Net… http://op… 7493e7… kml    "kml" 78ec5279-4… bcgeogr… FALSE           
-#> 
+#>   name       url                id         format ext   package_id     location  bcdata_available
+#>   <chr>      <chr>              <chr>      <chr>  <chr> <chr>          <chr>     <lgl>           
+#> 1 BC Geogra~ https://catalogue~ 529507c4-~ other  ""    78ec5279-4534~ bcgeogra~ FALSE           
+#> 2 WMS getCa~ https://openmaps.~ 57d8843b-~ wms    ""    78ec5279-4534~ bcgeogra~ TRUE            
+#> 3 KML Netwo~ http://openmaps.g~ 7493e749-~ kml    "kml" 78ec5279-4534~ bcgeogra~ FALSE           
 #> You can access the 'Resources' data frame using bcdc_tidy_resources()
 ```
 
@@ -79,25 +76,25 @@ This data is the boundary of each school district. The key information in this m
 ```r
 bcdc_query_geodata("78ec5279-4534-49a1-97e8-9d315936f08b")
 #> Querying 'school-districts-of-bc' record
[…]
