### 2021-03-31 2f197f5
grab entire capabilities xml and store in pkg env
