### 2021-03-31 fdd4107
use xml to feed into accessor fns
