+++ R/bcdc_options.R
@@ -18,7 +18,7 @@
 #' failed calls to the data catalogue. Options in R are reset every time R is re-started. See examples for
 #' addition ways to restore your initial state.
 #'
-#' `bcdata.max_geom_pred_size` is the maximum size of an object used for a geometric operation. Objects
+#' `bcdata.max_geom_pred_size` is the maximum size in bytes of an object used for a geometric operation. Objects
 #' that are bigger than this value will have a bounding box drawn and apply the geometric operation
 #' on that simpler polygon. Users can try to increase the maximum geometric predicate size and see
 #' if the bcdata catalogue accepts their request.
@@ -74,22 +74,21 @@ bcdc_options <- function() {
     "bcdata.chunk_limit", null_to_na(getOption("bcdata.chunk_limit")), 1000,
     "bcdata.single_download_limit",
     null_to_na(getOption("bcdata.single_download_limit",
-                         default = ._bcdataenv_$bcdata_dl_limit)), 10000
+                         default = bcdc_single_download_limit())), 10000
   )
 }
 
 
 check_chunk_limit <- function(){
   chunk_value <- getOption("bcdata.chunk_limit")
-  chunk_limit <- getOption("bcdata.single_download_limit", default = ._bcdataenv_$bcdata_dl_limit)
+  chunk_limit <- getOption("bcdata.single_download_limit", default = bcdc_single_download_limit())
 
   if(!is.null(chunk_value) && chunk_value >= chunk_limit){
     stop(glue::glue("Your chunk value of {chunk_value} exceed the BC Data Catalogue chunk limit of {chunk_limit}"), call. = FALSE)
   }
 }
 
-
-bcdc_single_download_limit <- function() {
+bcdc_get_capabilities_xml <- function() {
   if (has_internet()) {
     url <- "http://openmaps.gov.bc.ca/geo/pub/ows?service=WFS&version=2.0.0&request=Getcapabilities"
     cli <- bcdata:::bcdc_http_client(url, auth = FALSE)
@@ -101,8 +100,31 @@ bcdc_single_download_limit <- function() {
     ))
 
     res <- cc$parse("UTF-8")
-    doc <- xml2::read_xml(res)
+    xml2::read_xml(res)
 
+  } else {
+    message("No access to internet")
+    invisible(FALSE)
+  }
+}
+
+bcdc_get_capabilities <- function() {
+  doc <- ._bcdataenv_$get_capabilities_xml
+
+  features <- xml2::xml_child(doc, 4)
+  feature_list <- xml2::as_list(features)
+  purrr::map_dfr(feature_list, ~ {
+    list(whse_name = .x$Name[[1]] %||% NA_character_,
+         title = .x$Title[[1]] %||% NA_character_,
+         cat_url = attr(.x$MetadataURL, "href") %||% NA_character_)
+  })
+}
+
+
+bcdc_single_download_limit <- function() {
+  doc <- ._bcdataenv_$get_capabilities_xml
+
+  if (inherits(doc, "xml_document")) {
     constraints <- xml2::xml_find_all(doc, ".//ows:Constraint")
     count_defaults <- constraints[which(xml2::xml_attrs(constraints) %in% "CountDefault")]
     xml2::xml_double(count_defaults)
