### 2021-03-31 514d01e
add column descriptions and test with describe feature
