### 2021-03-31 bcgov/bcdata#246 @boshek
My sense is that the language on the rmarkdown issue template is stronger than the what we want. Yihue gets so many more requests will a _huge_ user base so I can understand maintaining judiciously structured issues. In our case, we still have a relatively small user base including many R beginners. We _want_ people to open issues even if they aren't perfectly formatted or have a complete reprex. We don't want to scare folks off. With that in mind I am thinking about this issue template which is taken almost verbatim from usethis:

```
---
name: Bug report or feature request
about: Describe a bug you've seen or make a case for a new feature
---

Please briefly describe your problem and what output you expect. Please include a link to the catalogue record that you are trying to access if applicable.

Please include a minimal reproducible example (AKA a reprex). If you've never heard of a [reprex](http://reprex.tidyverse.org/) before, start by reading <https://www.tidyverse.org/help/#reprex>.

Brief description of the problem

```r
