### 2021-04-01 ba112a1
Update R/cql-geom-predicates.R
Co-authored-by: Andy Teucher <andy.teucher@gov.bc.ca>
