### 2021-04-01 e123340
change to bcdc_check_geom_size
