+++ NAMESPACE
@@ -46,6 +46,7 @@ export(TOUCHES)
 export(WITHIN)
 export(as_tibble)
 export(bcdc_browse)
+export(bcdc_check_geom_size)
 export(bcdc_describe_feature)
 export(bcdc_get_data)
 export(bcdc_get_record)
@@ -58,7 +59,6 @@ export(bcdc_query_geodata)
 export(bcdc_read_functions)
 export(bcdc_search)
 export(bcdc_search_facets)
-export(bcdc_size_check)
 export(bcdc_tidy_resources)
 export(collect)
 export(filter)
