### 2021-04-01 a9030ba
return TRUE if check succeeds
