+++ tests/testthat/test-geom-operators.R
@@ -24,6 +24,7 @@ test_that("bcdc_check_geom_size outputs message with low threshold",{
 
   withr::local_options(list(bcdata.max_geom_pred_size = 1))
   expect_message(bcdc_check_geom_size(local))
+  expect_false(bcdc_check_geom_size(local))
 })
 
 test_that("bcdc_check_geom_size is silent with high threshold",{
@@ -31,7 +32,7 @@ test_that("bcdc_check_geom_size is silent with high threshold",{
   skip_if_net_down()
 
   withr::local_options(list(bcdata.max_geom_pred_size = 1E10))
-  expect_false(bcdc_check_geom_size(local))
+  expect_true(bcdc_check_geom_size(local))
 })
 
 
