### 2021-04-01 7c62700
Update R/bcdc_options.R
Co-authored-by: Andy Teucher <andy.teucher@gov.bc.ca>
