### 2021-04-01 24635e3
abstract out joining fn
