+++ NEWS.md
@@ -1,6 +1,9 @@
 # bcdata 0.2.2.9000
 ### IMPROVEMENTS
 - Setting the `bcdata.single_download_limit` limit dynamically from the getCapabilities endpoint. 
+- `bcdc_describe_feature` now joins a object description column to the returned object to provide more information about a field 
+directly in R. #241
+
 
 # bcdata 0.2.2
 ### IMPROVEMENTS
