### 2021-04-01 fdce293
finalized message about bounding box
