### 2021-04-01 6b5b5f0
pass predicate so it can be part of the message
