+++ vignettes/efficiently-query-spatial-data-in-the-bc-data-catalogue.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Querying Spatial Data with bcdata"
-date: "2021-03-31"
+date: "2021-04-01"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Querying Spatial Data with bcdata}
@@ -265,9 +265,10 @@ districts_parks <- bcdc_query_geodata("6a2fea1b-0cc4-4fc2-8017-eaf755d516da") %>
   collect()
 #> The object is too large to perform exact spatial operations using bcdata.
 #> Object size: 948576 bytes
-#> BC Data Threshold: 5e+05 bytes
+#> BC Data Threshold: 500000 bytes
 #> Exceedance: 448576 bytes
-#> See ?bcdc_size_check for more details
+#> See ?bcdc_check_geom_size for more details
+#> A bounding box was drawn around the object passed to INTERSECTS and all features within the box will be returned.
 ```
 
 Plotting both the filtered parks data and the district polygons reveals an important consideration when using `bcdata`:
@@ -278,12 +279,12 @@ In this example, many parks not contained within the school districts are includ
 
 
 ```r
-bcdc_size_check(districts)
+bcdc_check_geom_size(districts)
 #> The object is too large to perform exact spatial operations using bcdata.
 #> Object size: 948576 bytes
-#> BC Data Threshold: 5e+05 bytes
+#> BC Data Threshold: 500000 bytes
 #> Exceedance: 448576 bytes
-#> See ?bcdc_size_check for more details
+#> See ?bcdc_check_geom_size for more details
 ```
 
 Drawing the bounding box illustrates this point:
@@ -330,29 +331,29 @@ There are a couple of other functions in `bcdata` that are useful to know when w
 
 ```r
 bcdc_describe_feature("6a2fea1b-0cc4-4fc2-8017-eaf755d516da")
-#> # A tibble: 20 x 5
-#>    col_name      sticky remote_col_type    local_col_type column_comments                        
-#>    <chr>         <lgl>  <chr>              <chr>          <chr>                                  
-#>  1 id            FALSE  xsd:string         character      <NA>                                   
-#>  2 LOCAL_REG_GR~ FALSE  xsd:decimal        numeric        LOCAL_REG_GREENSPACE_ID is a system-ge~
-#>  3 PARK_NAME     TRUE   xsd:string         character      PARK NAME is the name of the park, e.g~
-#>  4 PARK_TYPE     TRUE   xsd:string         character      PARK_TYPE is the type of park, e.g., L~
-#>  5 PARK_PRIMARY~ TRUE   xsd:string         character      PARK PRIMARY USE defines the primary u~
-#>  6 REGIONAL_DIS~ TRUE   xsd:string         character      REGIONAL_DISTRICT is the name of the r~
-#>  7 MUNICIPALITY  TRUE   xsd:string         character      MUNICIPALITY is the name of the munici~
-#>  8 CIVIC_NUMBER  TRUE   xsd:decimal        numeric        CIVIC_NUMBER is the street number or n~
-#>  9 CIVIC_NUMBER~ TRUE   xsd:string         character      CIVIC_NUMBER_SUFFIX is the letter or f~
-#> 10 STREET_NAME   TRUE   xsd:string         character      STREET_NAME is the name of the street ~
-#> 11 LATITUDE      TRUE   xsd:decimal        numeric        LATITUDE is the geographic coordinate,~
-#> 12 LONGITUDE     TRUE   xsd:decimal        numeric        LONGITUDE is the geographic coordinate~
[…]
