### 2021-04-01 05728c3
turn precompile into a function closes #205
