### 2021-04-01 e78b379
bcdc_single_download_limit returns integer not double
