### 2021-04-01 7123846
Update R/cql-geom-predicates.R
Co-authored-by: Andy Teucher <andy.teucher@gov.bc.ca>
