### 2021-04-01 e5da445
bcdc_get_wfs_records deal with NULL xml
