### 2021-04-01 cfb5d99
Bbox text (#258)
* add bcdc_check_size fn

* add to ns

* add convenience cli fns

* specify units

* test for message instead of warning

* test size check with high and low

* update README

* show msg

* Update R/cql-geom-predicates.R

Co-authored-by: Andy Teucher <andy.teucher@gov.bc.ca>

* change to bcdc_check_geom_size

* replace ns and docs

* return TRUE if check succeeds

* ref vignette in docs

* small doc improvements

* update NEWS

* finalized message about bounding box

* pass predicate so it can be part of the message

* update vignette

* Update R/cql-geom-predicates.R

Co-auth
[…]
