### 2021-04-01 33eda68
test bcdc_get_wfs_records
