### 2021-04-02 11d2c5c
Add full XML get capabilities into pkg env to access object description (#259)
