+++ tests/testthat/test-options.R
@@ -34,7 +34,7 @@ test_that("bcdata.single_download_limit", {
 
 })
 
-test_that("bcdata.single_download_limit can be changes",{
+test_that("bcdata.single_download_limit can be changed",{
   withr::local_options(list(bcdata.single_download_limit = 13))
   expect_equal(getOption("bcdata.single_download_limit"), 13)
 })
@@ -43,5 +43,5 @@ test_that("bcdc_single_download_limit returns a number",{
   skip_on_cran()
   skip_if_net_down()
   lt <- bcdc_single_download_limit()
-  expect_type(lt, "double")
+  expect_type(lt, "integer")
 })
