### 2021-04-06 e45b0af
add issue template. Closes #246
