### 2021-04-07 b51f280
Address JOSS review issues #245 and #248
