+++ inst/JOSS/paper.md
@@ -15,7 +15,7 @@ affiliations:
     index: 1
   - name: Data Science Partnerships, Ministry of Citizens' Services, Province of British Columbia
     index: 2
-date: "2020-11-30"
+date: "2021-04-07"
 output:
   html_document:
     keep_md: yes
@@ -40,6 +40,10 @@ tags:
 # Introduction
 
 The British Columbia government hosts over 2000 tabular and geospatial data sets in the B.C. Data Catalogue (@bcdc).  Most provincial geospatial data is available through the B.C. Data Catalogue under an open licence, via a [Web Feature Service](https://en.wikipedia.org/wiki/Web_Feature_Service) (WFS). A Web Feature Service is a powerful and flexible service for distributing geographic features over the web, supporting both geospatial and non-spatial querying.  The `bcdata` package for the R programming language (@RCore) wraps two distinct but complimentary web APIs - one for the B.C. Data Catalogue and one for the Web Feature Service.  This allows R users to search, download and import metadata and data from the B.C. Data Catalogue, as well as efficiently query and directly read geospatial data from the Web Feature Service into their R session. The `bcdata` package implements a novel application of `dbplyr` (@dbplyr) using a Web Feature Service backend---rather than a database backend---where a locally constructed query is processed by a remote server. This allows for fast and efficient geospatial data retrieval using familiar `dplyr` syntax. Through this functionality the `bcdata` package connects British Columbia government open data holdings with the vast capabilities of R.
+
+# Related work
+
+Open data, and geospatial data science are currently popular topics in the R community. Packages related to `bcdata` include [ckanr](https://docs.ropensci.org/ckanr/) (@ckanr) for interacting with [CKAN](https://ckan.org/) instances, and [ows4R](https://github.com/eblondel/ows4R) (@ows4r) which provides a low-level [R6](https://CRAN.R-project.org/package=R6) interface to OWS/WFS servers. `bcdata` seamlessly unifies these operations for B.C. public data holdings, and provides a user-friendly interface using a functional programming style that is familiar to users of the popular `tidyverse` tools. There are many packages available for other jurisdictions' data portals (e.g., [opendatatoronto](https://sharlagelfand.github.io/opendatatoronto/), [opendataes](https://ropenspain.github.io/opendataes/index.html)) however as far as the authors are aware, no other packages provide the `dplyr` like interface to large spatial datasets via WFS.
 
 # Usage 
 
@@ -90,9 +94,9 @@ bcdc_tidy_resources("bc-schools-district-provincial-scholarships")
 # A tibble: 2 x 8
   name  id    format bcdata_available url   ext   package_id
   <chr> <chr> <chr>  <lgl>            <chr> <chr> <chr>     
-1 Awar~ 4e87~ xlsx   TRUE             http~ xlsx  651b60c2-~
-2 Awar~ 8a2c~ txt    TRUE             http~ txt   651b60c2-~
-# ... with 1 more variable: location <chr>
+1 Awar… 4e87… xlsx   TRUE             http… xlsx  651b60c2-…
+2 Awar… 8a2c… txt    TRUE             http… txt   651b60c2-…
+# … with 1 more variable: location <chr>
 ```
 
 
@@ -152,17 +156,17 @@ head(scholars)
 
 ```
 # A tibble: 6 x 9
-  SCHOOL_YEAR_ISS~ `Sub Pop Code` `Num Prov Schol~ `Num Prov Schol~
-  <chr>            <chr>          <chr>            <chr>           
-1 1996/1997        ALL STUDENTS   3509             20              
-2 1996/1997        FEMALE         1921             7               
-3 1996/1997        MALE           1588             13              
[…]
