### 2021-04-07 64af234
turn precompile into a function closes #205 (#260)
* turn precompile into a function closes #205

* comment out single vignette call
