### 2021-04-08 8258d43
Update inst/JOSS/paper.Rmd
Co-authored-by: Sam Albers <sam.albers@gov.bc.ca>
