### 2021-04-08 284d4bd
Add grepl shim, #227
