### 2021-04-09 fd73ed7
JOSS paper - change example from HA to Spp occurrences
