+++ inst/JOSS/paper.md
@@ -15,7 +15,7 @@ affiliations:
     index: 1
   - name: Data Science Partnerships, Ministry of Citizens' Services, Province of British Columbia
     index: 2
-date: "2021-04-07"
+date: "2021-04-09"
 output:
   html_document:
     keep_md: yes
@@ -43,7 +43,7 @@ The British Columbia government hosts over 2000 tabular and geospatial data sets
 
 # Related work
 
-Open data, and geospatial data science are currently popular topics in the R community. Packages related to `bcdata` include [ckanr](https://docs.ropensci.org/ckanr/) (@ckanr) for interacting with [CKAN](https://ckan.org/) instances, and [ows4R](https://github.com/eblondel/ows4R) (@ows4r) which provides a low-level [R6](https://CRAN.R-project.org/package=R6) interface to OWS/WFS servers. `bcdata` seamlessly unifies these operations for B.C. public data holdings, and provides a user-friendly interface using a functional programming style that is familiar to users of the popular `tidyverse` tools. There are many packages available for other jurisdictions' data portals (e.g., [opendatatoronto](https://sharlagelfand.github.io/opendatatoronto/), [opendataes](https://ropenspain.github.io/opendataes/index.html)) however as far as the authors are aware, no other packages provide the `dplyr` like interface to large spatial datasets via WFS.
+Open data, and geospatial data science are currently popular topics in the R community. Packages related to `bcdata` include [ckanr](https://docs.ropensci.org/ckanr/) (@ckanr) for interacting with [CKAN](https://ckan.org/) instances, and [ows4R](https://github.com/eblondel/ows4R) (@ows4r) which provides a low-level [R6](https://CRAN.R-project.org/package=R6) interface to OWS/WFS servers. `bcdata` seamlessly unifies these operations for B.C. public data holdings, and provides a user-friendly interface using a functional programming style that is familiar to users of the popular `tidyverse` tools. There are many packages available for other jurisdictions' data portals (e.g., [opendatatoronto](https://sharlagelfand.github.io/opendatatoronto/), [opendataes](https://ropenspain.github.io/opendataes/index.html)) however as far as the authors are aware, no other packages provide the `dplyr` like syntax to large spatial datasets via WFS.
 
 # Usage 
 
@@ -191,94 +191,96 @@ To demonstrate, we will query the Northern Health Authority boundary from the [H
 
 
 ```r
-## Get the metadata for the Health Authority Boundaries catalogue record
-ha_record <- bcdc_get_record("7bc6018f-bb4f-4e5d-845e-c529e3d1ac3b")
+## Get the metadata for the Species and Ecosystems at Risk Occurrences catalogue record
+sp_eco_record <- bcdc_get_record("0e035e55-f257-458f-9a96-80c01c69d389")
 
 ## Have a quick look at the geospatial columns to help with filter or select
-bcdc_describe_feature(ha_record)
+bcdc_describe_feature(sp_eco_record)
 ```
 
 ```
-# A tibble: 11 x 4
-   col_name            sticky remote_col_type          local_col_type
-   <chr>               <lgl>  <chr>                    <chr>         
- 1 id                  FALSE  xsd:string               character     
- 2 HLTH_HAB_SYSID      FALSE  xsd:decimal              numeric       
[…]
