### 2021-04-09 d665423
JOSS - tidy up sp-eco example
