### 2021-04-12 efb8076
create paginated only print msg
