### 2021-04-12 3383621
Update R/utils-classes.R
Co-authored-by: Andy Teucher <andy.teucher@gov.bc.ca>
