+++ R/utils-classes.R
@@ -395,7 +395,8 @@ collect.bcdc_promise <- function(x, ...){
     url <- cc$url
     full_url <- cli$url_fetch(query = query_list)
   } else {
-    message("This record requires pagination to complete the request.")
+    chunk <- getOption("bcdata.chunk_limit", default = 1000)
+    message(glue::glue("This record requires {ceiling(number_of_records/chunk)} paginated requests to complete."))
     sorting_col <- pagination_sort_col(x$cols_df)
 
     query_list <- c(query_list, sortby = sorting_col)
@@ -407,7 +408,7 @@ collect.bcdc_promise <- function(x, ...){
       limit_param = "count",
       offset_param = "startIndex",
       limit = number_of_records,
-      chunk = getOption("bcdata.chunk_limit", default = 1000),
+      chunk = chunk,
       progress = interactive()
     )
 
