### 2021-04-12 0f37bfc
Update R/utils-classes.R
Co-authored-by: Andy Teucher <andy.teucher@gov.bc.ca>
