### 2021-04-12 bcgov/bcdata#265 @ateucher
@boshek @stephhazlitt - re the wms/wfs/web service/web map service.  WFS (Web Feature Service) and WMS (Web Map Service) are both special cases of OWS (OGC Web Service). I think any catalogue record that has one of WFS of WMS will have both. The catalogue lists all of these records as WMS (which is why `bcdc_tidy_resources()` lists them as such and we use that [language throughout](https://github.com/bcgov/bcdata/search?q=wms)) though `bcdata` pretty much only makes use of the WFS (except in `bcdc_preview()` which is only just functional).

So the cadillac fix would be to go through it all and use ows/wms/wfs consitently where they are appropriate, but as a stopgap it might be worth just updating the documentation as Miles suggested. Thoughts?
