### 2021-04-12 bcgov/bcdata#265 @stephhazlitt
Could we just use Web Service throughout?
