### 2021-04-12 eac49f8
add statement of need + citations; closes #244 #243
