+++ inst/JOSS/paper.bib
@@ -10,6 +10,26 @@
   url = {https://data.gov.bc.ca/},
 }
 
+@Manual{tidyverse,
+    title = {Welcome to the {tidyverse}},
+    author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
+    year = {2019},
+    journal = {Journal of Open Source Software},
+    volume = {4},
+    number = {43},
+    pages = {1686},
+    doi = {10.21105/joss.01686},
+  }
+
+@Manual{tidydata,
+    title = {Tidy Data},
+    author = {Hadley Wickham},
+    year = {2014},
+    journal = {Journal of Statistical Software},
+    volume = {14},
+    doi = {10.18637/jss.v059.i10},
+    url = {https://www.jstatsoft.org/article/view/v059i10}
+  }
 
 @Manual{RCore,
   title = {R: A Language and Environment for Statistical Computing},
@@ -53,3 +73,16 @@ API},
     note = {R package version 0.6.0},
     url = {https://CRAN.R-project.org/package=ckanr},
   }
+
+@Manual{sf,
+    author = {Edzer Pebesma},
+    title = {{Simple Features for R: Standardized Support for Spatial Vector Data}},
+    year = {2018},
+    journal = {{The R Journal}},
+    doi = {10.32614/RJ-2018-009},
+    url = {https://doi.org/10.32614/RJ-2018-009},
+    pages = {439--446},
+    volume = {10},
+    number = {1},
+  }
+
