### 2021-04-13 95cf896
create paginated only print msg (#264)
