### 2021-04-13 bcgov/bcdata#265 @ateucher
I think "Web Map Service" might be better, since the catalogue is also a web service.
