### 2021-04-13 bcgov/bcdata#265 @ateucher
That said, the way you use "Web Feature Service" in #266 is correct, and switching that to "Web Map Service" would be wrong - WMS is the image overlay service, WFS is the actual vector shapes + attributes.
