### 2021-04-13 bcgov/bcdata#265 @ateucher
I say keep the article as you have it, and I'll try to align some of the documentation
