### 2021-04-13 f94eb3b
Update inst/JOSS/paper.Rmd
Co-authored-by: Sam Albers <sam.albers@gov.bc.ca>
