### 2021-04-13 ab82acc
Update inst/JOSS/paper.Rmd
Co-authored-by: Sam Albers <sam.albers@gov.bc.ca>
