### 2021-04-13 80d0fff
Update inst/JOSS/paper.Rmd
Co-authored-by: Sam Albers <sam.albers@gov.bc.ca>
