### 2021-04-13 4d15819
Fix docs in bcdc_query_geodata for wms/wfs - #265, JOSS UF3
