### 2021-04-13 e95eb2a
Unify language in vignettes for WFS/WMS - #265, JOSS UF3
