### 2021-04-13 c52400d
Fix WFS/WMS documentation throughout - #265, JOSS UF3
