+++ inst/JOSS/paper.bib
@@ -21,16 +21,6 @@
     doi = {10.21105/joss.01686},
   }
 
-@Manual{tidydata,
-    title = {Tidy Data},
-    author = {Hadley Wickham},
-    year = {2014},
-    journal = {Journal of Statistical Software},
-    volume = {14},
-    doi = {10.18637/jss.v059.i10},
-    url = {https://www.jstatsoft.org/article/view/v059i10}
-  }
-
 @Manual{RCore,
   title = {R: A Language and Environment for Statistical Computing},
   author = {{R Core Team}},
