### 2021-04-13 7be6259
bcdc_record print method tells user is WFS available via bcdcd_query_geodata
