### 2021-04-13 95ec537
tweak format of print.bcdc_record
