### 2021-04-13 908fa8a
dont need cask install for precompile action
