### 2021-04-13 43c4e9c
precompile vignettes
