+++ vignettes/bcdata.Rmd
@@ -71,8 +71,7 @@ Let's search the catalogue for records that contain the word "recycling":
 ## Give me the catalogue search results for 'recycling'
 bcdc_search("recycling")
 #> List of B.C. Data Catalogue Records
-#> 
-#> Number of records: 5
+#> Number of records: 4
 #> Titles:
 #> 1: BC FIRST Tire Recycling Data 1991-2006 (csv)
 #>  ID: a29ad492-29a2-44b9-8693-d27a8cc8e686
@@ -83,15 +82,12 @@ bcdc_search("recycling")
 #> 3: Environmental Protection Information Resources e-Library ()
 #>  ID: dae0f2c3-b4f4-4d16-a96d-d7fe7c1581f3
 #>  Name: environmental-protection-information-resources-e-library
-#> 4: Cross-linked Information Resources (CLIR) (other)
-#>  ID: 6f8576bc-1585-429c-97f9-e94b059fa455
-#>  Name: cross-linked-information-resources-clir
-#> 5: Ministry of Transportation (MOT) Surface Type (other, wms, kml)
+#> 4: Ministry of Transportation (MOT) Surface Type (other, wms, kml)
 #>  ID: 3a77f19f-2f49-416c-ad32-75e8d06e6c5a
-#>  Name: ministry-of-transportation-mot-surface-type 
+#>  Name: ministry-of-transportation-mot-surface-type
 #> 
-#> Access a single record by calling bcdc_get_record(ID)
-#>       with the ID from the desired record.
+#> Access a single record by calling `bcdc_get_record(ID)` with the ID
+#>  from the desired record.
 ```
 
 You can set the number of records to be returned from the search and/or you can customize your search using the catalogue search _facets_ `license_id`, `download_audience`, `type`, `res_format`, `sector`, and `organization`:
@@ -101,29 +97,27 @@ You can set the number of records to be returned from the search and/or you can 
 ## Give me the first catalogue search result for 'recycling'
 bcdc_search("recycling", n = 1)
 #> List of B.C. Data Catalogue Records
-#> 
 #> Number of records: 1
 #> Titles:
 #> 1: BC FIRST Tire Recycling Data 1991-2006 (csv)
 #>  ID: a29ad492-29a2-44b9-8693-d27a8cc8e686
-#>  Name: bc-first-tire-recycling-data-1991-2006 
+#>  Name: bc-first-tire-recycling-data-1991-2006
 #> 
-#> Access a single record by calling bcdc_get_record(ID)
-#>       with the ID from the desired record.
+#> Access a single record by calling `bcdc_get_record(ID)` with the ID
+#>  from the desired record.
 
 ## Give me the catalogue search results for 'recycling' where the
 ## data is tabular and licenced under Open Government Licence – British Columbia
 bcdc_search("recycling", type = "Dataset", license_id = "2")
 #> List of B.C. Data Catalogue Records
-#> 
 #> Number of records: 1
 #> Titles:
 #> 1: BC FIRST Tire Recycling Data 1991-2006 (csv)
 #>  ID: a29ad492-29a2-44b9-8693-d27a8cc8e686
-#>  Name: bc-first-tire-recycling-data-1991-2006 
+#>  Name: bc-first-tire-recycling-data-1991-2006
 #> 
-#> Access a single record by calling bcdc_get_record(ID)
-#>       with the ID from the desired record.
+#> Access a single record by calling `bcdc_get_record(ID)` with the ID
+#>  from the desired record.
 ```
 
 You can see all valid values for the catalogue search facets using `bcdata::bcdc_search_facets()`:
@@ -133,17 +127,18 @@ You can see all valid values for the catalogue search facets using `bcdata::bcdc
 ## Valid values for search facet 'license_id'
 bcdc_search_facets(facet = "license_id")
[…]
