+++ vignettes/explore-silviculture-data-using-bcdata.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Exploring Silviculture Data with bcdata"
-date: "2020-06-19"
+date: "2021-04-13"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Exploring Silviculture Data with bcdata}
@@ -62,58 +62,57 @@ We can gather the data we need to answer our question using the [RESULTS - silvi
 ```r
 # Get the metadata using the human-readable record name
 bcdc_get_record("results-forest-cover-silviculture")
-#> B.C. Data Catalogue Record: RESULTS - Forest Cover Silviculture 
-#> 
-#> Name: results-forest-cover-silviculture (ID: 258bb088-4113-47b1-b568-ce20bd64e3e3 )
-#> Permalink: https://catalogue.data.gov.bc.ca/dataset/258bb088-4113-47b1-b568-ce20bd64e3e3
+#> B.C. Data Catalogue Record: RESULTS - Forest Cover Silviculture
+#> Name: results-forest-cover-silviculture (ID:
+#>  258bb088-4113-47b1-b568-ce20bd64e3e3)
+#> Permalink:
+#>  https://catalogue.data.gov.bc.ca/dataset/258bb088-4113-47b1-b568-ce20bd64e3e3
 #> Sector: Natural Resources
 #> Licence: Open Government Licence - British Columbia
 #> Type: Geographic
-#> Last Updated: 2020-05-25 
-#> 
-#> Description: RESULTS opening's forest cover poylgons with silviculture component provided.
-#> Current forest cover subimssion into RESULTS must contain attribute and map
-#> information.  However, there are historical forest cover polygon infomration where
-#> maps are not available.  Forest Cover is provided at three critical milestones of at
-#> harvesting, at regeneration, and at free growing.  This is a part o fthe
-#> Silviculture and Land Status Tracking dataset, which includes tracking achievement
-#> of silviculture obligations on Crown Land 
-#> 
-#> Resources: (3)
-#> # A tibble: 3 x 8
-#>   name     url        id      format ext   package_id  location bcdata_available
-#>   <chr>    <chr>      <chr>   <chr>  <chr> <chr>       <chr>    <lgl>           
-#> 1 BC Geog… https://c… 92f845… other  ""    258bb088-4… bcgeogr… FALSE           
-#> 2 WMS get… https://o… 57670a… wms    ""    258bb088-4… bcgeogr… TRUE            
-#> 3 KML Net… https://o… d45a82… kml    "kml" 258bb088-4… bcgeogr… FALSE           
-#> 
-#> You can access the 'Resources' data frame using bcdc_tidy_resources()
+#> Description: RESULTS opening's forest cover poylgons with silviculture
+#>  component provided.  Current forest cover subimssion into RESULTS must
+#>  contain attribute and map information.  However, there are historical
+#>  forest cover polygon infomration where maps are not available.  Forest
+#>  Cover is provided at three critical milestones of at harvesting, at
+#>  regeneration, and at free growing.  This is a part o fthe Silviculture
+#>  and Land Status Tracking dataset, which includes tracking achievement
+#>  of silviculture obligations on Crown Land
+#> Available Resources (1):
+#>  1. WMS getCapabilities request (wms)
+#> Access the full 'Resources' data frame using:
+#>  bcdc_tidy_resources('258bb088-4113-47b1-b568-ce20bd64e3e3')
+#> Query and filter this data using:
+#>  bcdc_query_geodata('258bb088-4113-47b1-b568-ce20bd64e3e3')
 ```
 
[…]

[…]
