+++ R/bcdc-web-services.R
@@ -164,7 +164,8 @@ bcdc_query_geodata.bcdc_record <- function(record, crs = 3005) {
 #' Get preview map from the B.C. Web Map Service
 #'
 #' Note this does not return the actual map features, rather
-#' opens an image preview of the layer in [mapview::mapview()]
+#' opens an image preview of the layer in a
+#' [Leaflet](https://rstudio.github.io/leaflet/) map window
 #'
 #'
 #' @inheritParams bcdc_get_data
