### 2021-04-13 bcgov/bcdata#239 @ateucher
Summary of changes:

- [x] UF1: Issue #240, Closed by #264
- [x] UF2: Issue #241, Closed by #259
- [ ] UF3: Issue #265, Closed by #267
- [x] UF4: Unfixable, See https://github.com/bcgov/bcdata/issues/239#issuecomment-762187642
- [x]  UF5: Issue #243, Closed by #258
