### 2021-04-13 a0c3377
Vig width (#268)
* set width to 100

* directly specify where is_empty is coming from

* precompile vignettes

Co-authored-by: runner <runner@Mac-1618353632778.local>
