### 2021-04-13 f77599c
precompile vignettes
