+++ vignettes/explore-silviculture-data-using-bcdata.Rmd
@@ -63,3 +63,232 @@ We can gather the data we need to answer our question using the [RESULTS - silvi
 # Get the metadata using the human-readable record name
 bcdc_get_record("results-forest-cover-silviculture")
 #> B.C. Data Catalogue Record: RESULTS - Forest Cover Silviculture
+#> Name: results-forest-cover-silviculture (ID: 258bb088-4113-47b1-b568-ce20bd64e3e3)
+#> Permalink: https://catalogue.data.gov.bc.ca/dataset/258bb088-4113-47b1-b568-ce20bd64e3e3
+#> Sector: Natural Resources
+#> Licence: Open Government Licence - British Columbia
+#> Type: Geographic
+#> Description: RESULTS opening's forest cover poylgons with silviculture component
+#>  provided.  Current forest cover subimssion into RESULTS must contain attribute and map
+#>  information.  However, there are historical forest cover polygon infomration where maps
+#>  are not available.  Forest Cover is provided at three critical milestones of at
+#>  harvesting, at regeneration, and at free growing.  This is a part o fthe Silviculture
+#>  and Land Status Tracking dataset, which includes tracking achievement of silviculture
+#>  obligations on Crown Land
+#> Available Resources (1):
+#>  1. WMS getCapabilities request (wms)
+#> Access the full 'Resources' data frame using:
+#>  bcdc_tidy_resources('258bb088-4113-47b1-b568-ce20bd64e3e3')
+#> Query and filter this data using:
+#>  bcdc_query_geodata('258bb088-4113-47b1-b568-ce20bd64e3e3')
+```
+
+We see that this is a [Web Feature Service-enabled](https://en.wikipedia.org/wiki/Web_Feature_Service) geospatial data set--the list of data resources includes `WMS getCapabilities request`--so we can query and retrieve this geospatial data set using `bcdc_query_geodata()`:
+
+
+```r
+# Query the data using the permanent ID of the record to guard against name changes
+bcdc_query_geodata("258bb088-4113-47b1-b568-ce20bd64e3e3")
+#> Querying 'results-forest-cover-silviculture' record
+#> ● Using collect() on this object will return 888963 features and 159 fields
+#> ● Accessing this record requires pagination and will make 889 separate requests to the WFS.
+#> ● See ?bcdc_options
+#> ● At most six rows of the record are printed here
+#> ────────────────────────────────────────────────────────────────────────────────────────────────────
+#> Simple feature collection with 6 features and 159 fields
+#> Geometry type: MULTIPOLYGON
+#> Dimension:     XY
+#> Bounding box:  xmin: 1184167 ymin: 676633.7 xmax: 1411085 ymax: 1083545
+#> Projected CRS: NAD83 / BC Albers
+#> # A tibble: 6 x 160
+#>   id                 FOREST_COVER_ID STOCKING_STANDARD… OPENING_ID STANDARDS_UNIT_… SILV_POLYGON_NU…
+#>   <chr>                        <int>              <int>      <int> <chr>            <chr>           
+#> 1 WHSE_FOREST_VEGET…         3994018            1805957    1486231 2                2               
[…]
