+++ vignettes/efficiently-query-spatial-data-in-the-bc-data-catalogue.Rmd
@@ -53,20 +53,17 @@ bcdc_get_record("78ec5279-4534-49a1-97e8-9d315936f08b")
 #> Sector: Education
 #> Licence: Open Government Licence - British Columbia
 #> Type: Geographic
-#> Last Updated: 2020-11-03
 #> Description: The School Districts dataset contains the spatial representation (polygon)
 #>  of the current extent of the administrative areas defined under section 176(1) of the
 #>  School Act for the purposes of preservation and promotion of the fundamental principle
 #>  of local autonomy and control of public education at the public and governmental levels
 #>  through locally elected school boards.
-#> Resources: (3)
-#> # A tibble: 3 x 8
-#>   name        url                  id        format ext   package_id     location   bcdata_available
-#>   <chr>       <chr>                <chr>     <chr>  <chr> <chr>          <chr>      <lgl>           
-#> 1 BC Geograp… https://catalogue.d… 529507c4… other  ""    78ec5279-4534… bcgeograp… FALSE           
-#> 2 WMS getCap… https://openmaps.go… 57d8843b… wms    ""    78ec5279-4534… bcgeograp… TRUE            
-#> 3 KML Networ… http://openmaps.gov… 7493e749… kml    "kml" 78ec5279-4534… bcgeograp… FALSE           
-#> You can access the 'Resources' data frame using bcdc_tidy_resources()
+#> Available Resources (1):
+#>  1. WMS getCapabilities request (wms)
+#> Access the full 'Resources' data frame using:
+#>  bcdc_tidy_resources('78ec5279-4534-49a1-97e8-9d315936f08b')
+#> Query and filter this data using:
+#>  bcdc_query_geodata('78ec5279-4534-49a1-97e8-9d315936f08b')
 ```
 
 This data is the boundary of each school district. The key information in this metadata is that the layer has a resource in `"wms"` format ---which means it is available through a Web Feature Service. From this we know we can make use of `bcdc_query_geodata`.
@@ -174,7 +171,6 @@ bcdc_get_record("6a2fea1b-0cc4-4fc2-8017-eaf755d516da")
 #> Sector: Natural Resources
 #> Licence: Open Government Licence - British Columbia
 #> Type: Geographic
-#> Last Updated: 2020-11-25
 #> Description: This dataset contains spatial and attribute information for local and
 #>  regional greenspaces in British Columbia. Local and regional greenspaces are municipal
 #>  or regional district lands designated by local government agencies and managed for
@@ -183,15 +179,13 @@ bcdc_get_record("6a2fea1b-0cc4-4fc2-8017-eaf755d516da")
 #>  Open Government Licence, and in other cases, publicize parks and greenspaces on web maps
 #>  or pdf maps. Boundaries were edge-matched to the ParcelMap BC cadastre.  This spatial
 #>  layer contains multipart polygons.
-#> Resources: (4)
-#> # A tibble: 4 x 8
-#>   name        url                  id        format ext   package_id     location   bcdata_available
-#>   <chr>       <chr>                <chr>     <chr>  <chr> <chr>          <chr>      <lgl>           
-#> 1 WMS getCap… https://openmaps.go… a8d5b53e… wms    ""    6a2fea1b-0cc4… bcgeograp… TRUE            
[…]
