### 2021-04-14 a0053aa
Fix size comparison, render paper.md
