### 2021-04-14 bcgov/bcdata#239 @ateucher
All user feedback issues now addressed
