### 2021-04-14 #2927 @ateucher
@MilesMcBain's review identified several blockers to acceptance:

 - bcgov/bcdata#244: Addressed by bcgov/bcdata#266
 - bcgov/bcdata#246: Addressed by bcgov/bcdata#261
 - bcgov/bcdata#245: Addressed by bcgov/bcdata#263
 - bcgov/bcdata#242: Addressed by bcgov/bcdata#266

He also had one issue which he strongly recommended we address in terms of user-experience:

- bcgov/bcdata#243: Addressed by bcgov/bcdata#258

@MilesMcBain also had several other [excellent suggestions](https://github.com/bcgov/bcdata/issues/239) that would make the package better and more consistent for users, and we did address all of those that were possible:

- UF1: bcgov/bcdata#240, Closed by bcgov/bcdata#264
- UF2: bcgov/bcdata#241, Closed by bcgov/bcdata#259
- UF3: bcgov/bcdata#265, Closed by bcgov/bcdata#267
- UF4: Unfixable, See this [comment](https://github.com/bcgov/bcdata#239#issuecomment-762187642)

@tweed1e had [one blocker to acceptance](https://github.com/bcgov/bcdata/issues/248), which was addressed by bcgov/bcdata#263

Overall the review process was excellent and has made the package, and the paper, much better. Thank you very much to both reviewers!
