+++ NEWS.md
@@ -1,4 +1,5 @@
-# bcdata 0.2.2.9000
+# bcdata 0.2.3
+
 ### IMPROVEMENTS
 - Setting the `bcdata.single_download_limit` limit dynamically from the getCapabilities endpoint. #256
 - `bcdc_describe_feature` now joins an object description column to the returned object to provide more information about a field directly in R. #241, #259
