+++ cran-comments.md
@@ -1,7 +1,9 @@
 ### CRAN check issues
+
 * Fixed an issue where a test that called a web resource would sporadically cause a check failure if the resource was unavailable.
 
 ## Test environments
+
 * local macOS install (macOS Mojave 10.14.6), R 4.0.4
 * local Windows 10 install, R 4.0.5
 * ubuntu 18.04 (on github actions), R 4.0.5, 3.6.3, R 3.5
@@ -13,3 +15,10 @@
 ## R CMD check results
 
 0 errors | 0 warnings | 0 notes
+
+## Reverse dependencies
+
+We checked 1 reverse dependency, comparing R CMD check results across CRAN and dev versions of this package.
+
+We saw 0 new problems
+We failed to check 0 packages
