+++ revdep/README.md
@@ -1,0 +1,98 @@
+# Platform
+
+|field    |value                        |
+|:--------|:----------------------------|
+|version  |R version 4.0.4 (2021-02-15) |
+|os       |macOS Mojave 10.14.6         |
+|system   |x86_64, darwin17.0           |
+|ui       |X11                          |
+|language |(EN)                         |
+|collate  |en_CA.UTF-8                  |
+|ctype    |en_CA.UTF-8                  |
+|tz       |America/Vancouver            |
+|date     |2021-04-23                   |
+
+# Dependencies
+
+|package           |old      |new      |Δ  |
+|:-----------------|:--------|:--------|:--|
+|bcdata            |0.2.2    |0.2.3    |*  |
+|assertthat        |0.2.1    |0.2.1    |   |
+|base64enc         |0.1-3    |0.1-3    |   |
+|BH                |1.75.0-0 |1.75.0-0 |   |
+|blob              |1.2.1    |1.2.1    |   |
+|cellranger        |1.1.0    |1.1.0    |   |
+|classInt          |0.4-3    |0.4-3    |   |
+|cli               |2.4.0    |2.4.0    |   |
+|clipr             |0.7.1    |0.7.1    |   |
+|colorspace        |2.0-0    |2.0-0    |   |
+|cpp11             |0.2.7    |0.2.7    |   |
+|crayon            |1.4.1    |1.4.1    |   |
+|crosstalk         |1.1.1    |1.1.1    |   |
+|crul              |1.1.0    |1.1.0    |   |
+|curl              |4.3      |4.3      |   |
+|DBI               |1.1.1    |1.1.1    |   |
+|dbplyr            |2.1.1    |2.1.1    |   |
+|digest            |0.6.27   |0.6.27   |   |
+|dplyr             |1.0.5    |1.0.5    |   |
+|e1071             |1.7-6    |1.7-6    |   |
+|ellipsis          |0.3.1    |0.3.1    |   |
+|fansi             |0.4.2    |0.4.2    |   |
+|farver            |2.1.0    |2.1.0    |   |
+|generics          |0.1.0    |0.1.0    |   |
+|ggplot2           |3.3.3    |3.3.3    |   |
+|glue              |1.4.2    |1.4.2    |   |
+|gridExtra         |2.3      |2.3      |   |
+|gtable            |0.3.0    |0.3.0    |   |
+|hms               |1.0.0    |1.0.0    |   |
+|htmltools         |0.5.1.1  |0.5.1.1  |   |
+|htmlwidgets       |1.5.3    |1.5.3    |   |
+|httpcode          |0.3.0    |0.3.0    |   |
+|isoband           |0.2.4    |0.2.4    |   |
+|jsonlite          |1.7.2    |1.7.2    |   |
+|labeling          |0.4.2    |0.4.2    |   |
+|lazyeval          |0.2.2    |0.2.2    |   |
+|leaflet           |2.0.4.1  |2.0.4.1  |   |
+|leaflet.extras    |1.0.0    |1.0.0    |   |
+|leaflet.providers |1.9.0    |1.9.0    |   |
+|lifecycle         |1.0.0    |1.0.0    |   |
+|magrittr          |2.0.1    |2.0.1    |   |
+|markdown          |1.1      |1.1      |   |
+|mime              |0.10     |0.10     |   |
+|munsell           |0.5.0    |0.5.0    |   |
+|pillar            |1.6.0    |1.6.0    |   |
+|pkgconfig         |2.0.3    |2.0.3    |   |
+|png               |0.1-7    |0.1-7    |   |
+|prettyunits       |1.1.1    |1.1.1    |   |
[…]
