### 2021-04-26 c378bc9
Remove bcdc_get_wfs_records_xml from onLoad
- set get_capabilities_xml to NULL on pkg load
- bcdc_get_wfs_records_xml() retrieves xml on first request and store for rest of the session
- eliminate some redundant messages
