### 2021-04-26 6b27b25
Wrap example in try(), add missing skip_if_net_down()
