+++ NEWS.md
@@ -1,3 +1,5 @@
+# bcdata (development version)
+
 # bcdata 0.2.3
 
 ### IMPROVEMENTS
