### 2021-04-26 0f25da4
Remove redundant message, add to bcdc_single_download_limit
