### 2021-04-26 d310236
change name bcdc_get_wfs_records_xml -> bcdc_get_capabilities
