### 2021-04-26 0eeabd3
update NEWS.md
