### 2021-04-27 8b15290
Only run tests if testthat is available as it's only a suggested package
