### 2021-04-28 2fba4cd
Update cran-comments for v 0.2.4
