### 2021-04-28 51f035d
use try on getCapabilities request
