### 2021-04-28 3f3f111
Remove redundant message - downstream functions will emit appropriate messaging
