+++ NEWS.md
@@ -1,4 +1,4 @@
-# bcdata (development version)
+# bcdata 0.2.4
 
 * Code in `.onLoad()` that sent a request to the wfs getCapabilities endpoint could cause the package to fail to load. This was moved into an internal function `bcdc_get_capabilities()` that makes the request the first time it's required, and stores the result for the remainder of the session (#271)
 * testthat is now used conditionally to only run tests if the testthat package is installed.
