+++ cran-comments.md
@@ -9,16 +9,16 @@
 * local macOS install (macOS Mojave 10.14.6), R 4.0.4
 * local Windows 10 install, R 4.0.5
 * ubuntu 18.04 (on github actions), R 4.0.5, 3.6.3, R 3.5
-* ubuntu 18.04 (on github actions), R-devel (2021-04-27 r80234)
+* ubuntu 18.04 (on github actions), R-devel (2021-04-26 r80229)
 * Windows Server 2019 (on github actions), R 4.0.5
 * macOS Catalina 10.15 (on github actions), R 4.0.5
-* win-builder (devel: R version 4.1.0 alpha (2021-04-22 r80209))
+* win-builder (devel: R version 4.1.0 alpha (2021-04-28 r80240))
 
 ## R CMD check results
 
 0 errors | 0 warnings | 1 note
 
-There was one NOTE: "Days since last update: 4" - as stated above, I believe the fixes in this release warrant such a quick update.
+There was one NOTE: "Days since last update: 6" - as stated above, we believe the fixes in this release warrant such a quick update.
 
 ## Reverse dependencies
 
