### 2021-05-03 90afcef
Skip bcdc_options tests on cran as they do try to access internet
