### 2021-05-03 537e8a4
update cran-comments for resubmission
