### 2021-05-04 60cbbce
update cran-comments for resubmission
