### 2021-05-04 861e256
update cran-comments for resubmission
