+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: bcdata
 Title: Search and Retrieve Data from the BC Data Catalogue
-Version: 0.2.4
+Version: 0.2.4.9000
 Authors@R: 
     c(person(given = "Andy",
              family = "Teucher",
