+++ inst/JOSS/paper.Rmd
@@ -33,6 +33,7 @@ tags:
 knitr::opts_chunk$set(echo = TRUE, warning = FALSE, 
                       message = FALSE, fig.width = 8, 
                       fig.height = 12, fig.path = "",
+                      dev="ragg_png",
                       comment = NA)
 ```
 
